file committed
I implemented edge detection algorithms in MATLAB and conducted disease detection on fruits using the Fruit360 dataset. This involved analyzing the dataset, applying the algorithms, and identifying potential diseases based on the patterns recognized. This work contributes to the field of automated disease detection in fruits, potentially aiding in early disease identification and treatment.